streamlit-image-classification/requirements.txt
requests
pillow
numpy
matplotlib
colour
plotly
streamlit
streamlit-echarts